Split files, handle API errors safely, use proper serverless-offline package
