+++ serverless.yml
@@ -3,7 +3,7 @@ service: lockbot
 plugins:
   - serverless-webpack
   - serverless-dynamodb-local
-  - "@connorads/serverless-offline"
+  - serverless-offline
 provider:
   name: aws
   runtime: nodejs12.x
@@ -38,7 +38,7 @@ provider:
 
 functions:
   slack-handler:
-    handler: src/handlers/slack.handler
+    handler: src/handlers/slack/index.handler
     events:
       - http:
           method: post
@@ -50,7 +50,7 @@ functions:
           method: get
           path: /slack/oauth_redirect
   api-handler:
-    handler: src/handlers/api.handler
+    handler: src/handlers/api/index.handler
     events:
       - http:
           method: get
